--- data/src/main/java/com/tht/tht/data/remote/datasource/Datasource.kt
@@ -1,3 +1,0 @@
-package com.tht.tht.data.remote.datasource
-
-// 데이터소스 구현 클래스
